auto-ds-backend/requirements_base.txt
django==4.2.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
django-allauth==0.54.0
django-import-export==3.3.1
dj-rest-auth==4.0.1